Change name instrument registry to task registry
