+++ pyproject.toml
@@ -1,18 +1,18 @@
 [tool.poetry]
-name = "instrument-registry"
+name = "task-registry"
 version = "0.1.0"
 readme = "README.md"
 description = ""
 authors = ["ai-validatie-team <ai-validatie@minbzk.nl>"]
-repository = "https://github.com/MinBZK/instrument-registry"
-keywords = ["AI", "Validation", "Instrument", "Registry"]
+repository = "https://github.com/MinBZK/task-registry"
+keywords = ["AI", "Validation", "Instrument", "Task", "Registry"]
 license = "EUPL-1.2"
 packages = [
-    { include = "instrument_registry" }
+    { include = "task_registry" }
 ]
 
 [tool.poetry.scripts]
-deploy = "instrument_registry.data:main"
+deploy = "task_registry.data:main"
 
 [tool.poetry.dependencies]
 python = "^3.12"
@@ -43,13 +43,13 @@ types-pyyaml = "^6.0.12.20240724"
 
 
 [tool.poetry.urls]
-"Issue Tracker" = "https://github.com/MinBZK/instrument-registry/issues"
+"Issue Tracker" = "https://github.com/MinBZK/task-registry/issues"
 
 # Ruff settings: https://docs.astral.sh/ruff/configuration/
 [tool.ruff]
 line-length = 120
 target-version = "py311"
-src = ["instrument_registry", "tests", "script"]
+src = ["task_registry", "tests", "script"]
 include = ['script/validate']
 
 [tool.ruff.format]
@@ -91,7 +91,7 @@ fail_under = 95
 
 [tool.coverage.html]
 directory = "htmlcov"
-title = "instrument-registry"
+title = "task-registry"
 
 
 [tool.pytest.ini_options]
