+++ sonar-project.properties
@@ -1,13 +1,13 @@
-sonar.projectKey=MinBZK_instrument-registry
+sonar.projectKey=MinBZK_task-registry
 sonar.organization=minbzk
 
 # This is the name and version displayed in the SonarCloud UI.
-sonar.projectName=instrument_registry
+sonar.projectName=task_registry
 sonar.projectVersion=0.1.0
 
 
 # Path is relative to the sonar-project.properties file. Replace "\" by "/" on Windows.
-sonar.sources=instrument_registry/
+sonar.sources=task_registry/
 sonar.tests=tests/
 
 # Encoding of the source code. Default is default system encoding
