+++ task_registry/core/config.py
@@ -1,8 +1,8 @@
 import logging
 from typing import Any, TypeVar
 
-from instrument_registry.core.types import EnvironmentType, LoggingLevelType
 from pydantic_settings import BaseSettings
+from task_registry.core.types import EnvironmentType, LoggingLevelType
 
 logger = logging.getLogger(__name__)
 
@@ -10,8 +10,8 @@ logger = logging.getLogger(__name__)
 # Self type is not available in Python 3.10 so create our own with TypeVar
 SelfSettings = TypeVar("SelfSettings", bound="Settings")
 
-PROJECT_NAME: str = "IR"
-PROJECT_DESCRIPTION: str = "Instrument Registry"
+PROJECT_NAME: str = "TR"
+PROJECT_DESCRIPTION: str = "Task Registry"
 VERSION: str = "0.1.0"  # replace in CI/CD pipeline
